Move Redux store management to a separate file
This cleans up the ReduxService file a touch, leaving it focused on
coordinating services and the store.
