+++ background/services/redux/index.ts
@@ -1,15 +1,8 @@
-import browser, { runtime } from "webextension-polyfill"
-import { alias, wrapStore } from "webext-redux"
-import { configureStore, isPlain, Middleware } from "@reduxjs/toolkit"
-import { devToolsEnhancer } from "@redux-devtools/remote"
+import { runtime } from "webextension-polyfill"
 import { PermissionRequest } from "@tallyho/provider-bridge-shared"
-import { debounce } from "lodash"
 import { utils } from "ethers"
 
-import { diff as deepDiff } from "./differ"
 import {
-  decodeJSON,
-  encodeJSON,
   getEthereumNetwork,
   isProbablyEVMAddress,
   normalizeEVMAddress,
@@ -43,7 +36,6 @@ import { SignedTransaction } from "../../networks"
 import { AccountBalance, AddressOnNetwork, NameOnNetwork } from "../../accounts"
 import { Eligible, ReferrerStats } from "../island/types"
 
-import rootReducer from "../../redux-slices"
 import {
   AccountType,
   deleteAccount,
@@ -101,7 +93,6 @@ import {
   updateRollupEstimates,
 } from "../../redux-slices/transaction-construction"
 import { selectDefaultNetworkFeeSettings } from "../../redux-slices/selectors/transactionConstructionSelectors"
-import { allAliases } from "../../redux-slices/utils"
 import {
   requestPermission,
   emitter as providerBridgeSliceEmitter,
@@ -137,8 +128,8 @@ import {
   ETHEREUM,
   FLASHBOTS_RPC_URL,
   OPTIMISM,
-  USD,
   POPUP_MONITOR_PORT_NAME,
+  USD,
 } from "../../constants"
 import {
   clearApprovalInProgress,
@@ -149,10 +140,6 @@ import {
   SignatureResponse,
   TXSignatureResponse,
 } from "../signing"
-import {
-  migrateReduxState,
-  REDUX_STATE_VERSION,
-} from "../../redux-slices/migrations"
 import { PermissionMap } from "../provider-bridge/utils"
 import { TAHO_INTERNAL_ORIGIN } from "../internal-ethereum-provider/constants"
 import {
@@ -209,98 +196,9 @@ import { makeFlashbotsProviderCreator } from "../chain/serial-fallback-provider"
 import { AnalyticsPreferences, DismissableItem } from "../preferences"
 import { newPricePoints } from "../../redux-slices/prices"
 import NotificationsService from "../notifications"
+import { ReduxStoreType, initializeStore, readAndMigrateState } from "./store"
 
-// This sanitizer runs on store and action data before serializing for remote
-// redux devtools. The goal is to end up with an object that is directly
-// JSON-serializable and deserializable; the remote end will display the
-// resulting objects without additional processing or decoding logic.
-const devToolsSanitizer = (input: unknown) => {
-  switch (typeof input) {
-    // We can make use of encodeJSON instead of recursively looping through
-    // the input
-    case "bigint":
-    case "object":
-      return JSON.parse(encodeJSON(input))
-    // We only need to sanitize bigints and objects that may or may not contain
-    // them.
-    default:
-      return input
-  }
-}
-
-const persistStoreFn = <T>(state: T) => {
-  if (process.env.WRITE_REDUX_CACHE === "true") {
-    // Browser extension storage supports JSON natively, despite that we have
-    // to stringify to preserve BigInts
-    browser.storage.local.set({
-      state: encodeJSON(state),
-      version: REDUX_STATE_VERSION,
-    })
-  }
-}
-
-const persistStoreState = debounce(persistStoreFn, 50, {
-  trailing: true,
-  maxWait: 50,
-})
-
-const reduxCache: Middleware = (store) => (next) => (action) => {
-  const result = next(action)
-  const state = store.getState()
-
-  persistStoreState(state)
-  return result
-}
-
-// Declared out here so ReduxStoreType can be used in Main.store type
-// declaration.
-const initializeStore = (main: Main, preloadedState: object) =>
-  configureStore({
-    preloadedState,
-    reducer: rootReducer,
-    middleware: (getDefaultMiddleware) => {
-      const middleware = getDefaultMiddleware({
-        serializableCheck: {
-          isSerializable: (value: unknown) =>
-            isPlain(value) || typeof value === "bigint",
-        },
-        thunk: { extraArgument: { main } },
-      })
-
-      // It might be tempting to use an array with `...` destructuring, but
-      // unfortunately this fails to preserve important type information from
-      // `getDefaultMiddleware`. `push` and `pull` preserve the type
-      // information in `getDefaultMiddleware`, including adjustments to the
-      // dispatch function type, but as a tradeoff nothing added this way can
-      // further modify the type signature. For now, that's fine, as these
-      // middlewares don't change acceptable dispatch types.
-      //
-      // Process aliases before all other middleware, and cache the redux store
-      // after all middleware gets a chance to run.
-      middleware.unshift(alias(allAliases))
-      middleware.push(reduxCache)
-
-      return middleware
-    },
-    devTools: false,
-    enhancers:
-      process.env.NODE_ENV === "development"
-        ? [
-            devToolsEnhancer({
-              hostname: "localhost",
-              port: 8000,
-              realtime: true,
-              actionSanitizer: devToolsSanitizer,
-              stateSanitizer: devToolsSanitizer,
-            }),
-          ]
-        : [],
-  })
-
-type ReduxStoreType = ReturnType<typeof initializeStore>
-
-// TODO Rename ReduxService or CoordinationService, move to services/, etc.
-export default class Main extends BaseService<never> {
+export default class ReduxService extends BaseService<never> {
   /**
    * The redux store for the wallet core. Note that the redux store is used to
    * render the UI (via webext-redux), but it is _not_ the source of truth.
@@ -310,7 +208,7 @@ export default class Main extends BaseService<never> {
    */
   store: ReduxStoreType
 
-  static create: ServiceCreatorFunction<never, Main, []> = async () => {
+  static create: ServiceCreatorFunction<never, ReduxService, []> = async () => {
     const preferenceService = PreferenceService.create()
 
     const internalSignerService =
@@ -372,37 +270,10 @@ export default class Main extends BaseService<never> {
         )
       : getNoopService<WalletConnectService>()
 
-    let savedReduxState = {}
-    // Setting READ_REDUX_CACHE to false will start the extension with an empty
-    // initial state, which can be useful for development
-    if (process.env.READ_REDUX_CACHE === "true") {
-      const { state, version } = await browser.storage.local.get([
-        "state",
-        "version",
-      ])
-
-      if (state) {
-        const restoredState = decodeJSON(state)
-
-        if (typeof restoredState === "object" && restoredState !== null) {
-          // If someone managed to sneak JSON that decodes to typeof "object"
-          // but isn't a Record<string, unknown>, there is a very large
-          // problem...
-          savedReduxState = migrateReduxState(
-            restoredState as Record<string, unknown>,
-            version || undefined,
-          )
-        } else {
-          throw new Error(`Unexpected JSON persisted for state: ${state}`)
-        }
-      } else {
-        // Should be false if you don't want new users to see the modal
-        window.localStorage.setItem("modal_meet_taho", "false")
-      }
-    }
+    const savedReduxState = readAndMigrateState()
 
     return new this(
-      savedReduxState,
+      await savedReduxState,
       await preferenceService,
       await chainService,
       await enrichmentService,
@@ -528,83 +399,6 @@ export default class Main extends BaseService<never> {
 
     // Start up the redux store and set it up for proxying.
     this.store = initializeStore(this, savedReduxState)
-
-    /**
-     * Tracks pending updates to the redux store. This is used to delay responding
-     * to dispatched thunks until the store has been updated in the UI. This is
-     * necessary to prevent race conditions where the UI expects the store to be
-     * updated before the thunk has finished dispatching.
-     */
-    let storeUpdateLock: Promise<void> | null
-    let releaseLock: () => void
-
-    const queueUpdate = debounce(
-      (lastState, newState, updateFn) => {
-        if (lastState !== newState) {
-          const diff = deepDiff(lastState, newState)
-
-          if (diff !== undefined) {
-            updateFn(newState, [diff])
-          }
-        }
-        releaseLock()
-      },
-      30,
-      { maxWait: 30, trailing: true },
-    )
-
-    wrapStore(this.store, {
-      serializer: encodeJSON,
-      deserializer: decodeJSON,
-      diffStrategy: (oldObj, newObj, forceUpdate) => {
-        if (!storeUpdateLock) {
-          storeUpdateLock = new Promise((resolve) => {
-            releaseLock = () => {
-              resolve()
-              storeUpdateLock = null
-            }
-          })
-        }
-
-        queueUpdate(oldObj, newObj, forceUpdate)
-
-        // Return no diffs as we're manually handling these inside `queueUpdate`
-        return []
-      },
-      dispatchResponder: async (
-        dispatchResult: Promise<unknown> | unknown,
-        send: (param: { error: string | null; value: unknown | null }) => void,
-      ) => {
-        try {
-          // if dispatch is a thunk, wait for the result
-          const result = await dispatchResult
-
-          // By this time, all pending updates should've been tracked.
-          // since we're dealing with a thunk, we need to wait for
-          // the store to be updated
-          await storeUpdateLock
-
-          send({
-            error: null,
-            value: encodeJSON(result),
-          })
-        } catch (error) {
-          logger.error(
-            "Error awaiting and dispatching redux store result: ",
-            error,
-          )
-
-          // Store could still have been updated if there was an error
-          await storeUpdateLock
-
-          send({
-            error: encodeJSON(error),
-            value: null,
-          })
-        }
-      },
-    })
-
     this.initializeRedux()
   }
